Code corrections | Option 9 DONE! | Started option 10

+++ ProjetoPROG2/ProjetoPROG2/Client.h
@@ -18,7 +18,7 @@ class Client{
   unsigned short familySize;  // number of family members
   Address address; // client's address  
   vector<Packet*> packets; // vector to store client's packets bought
-  unsigned totalPurchased; // total value spent by the client
+  double totalPurchased; // total value spent by the client
 
  public:
   Client(); // Default Constructor
@@ -32,7 +32,7 @@ class Client{
   unsigned short getFamilySize() const;
   Address getAddress() const;
   vector<Packet*>& getPacketList();
-  unsigned getTotalPurchased() const;
+  double getTotalPurchased() const;
   
   // SET methods
   
@@ -41,7 +41,7 @@ class Client{
   void setFamilySize(unsigned short familySize);
   void setAddress(Address address);
   int setPacketList(string ids, vector<Packet> &packets);
-  void setTotalPurchased(unsigned totalPurchased);
+  void setTotalPurchased(double totalPurchased);
   
   // other methods
   friend ostream& operator<<(ostream& out, const Client & client);
